[Refactor] chatProf 구조 재정리(#61)
